+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "node ./bin/www"
+    "start": "node ./server.js"
   },
   "dependencies": {
     "bitfinex-api-node": "^1.0.2",
@@ -12,8 +12,8 @@
     "debug": "~2.2.0",
     "dotenv": "^4.0.0",
     "express": "~4.13.4",
-    "hbs": "~4.0.0",
     "morgan": "~1.7.0",
+    "pug": "^2.0.0-rc.4",
     "serve-favicon": "~2.3.0"
   }
 }
