+++ public/stylesheets/style.css
@@ -1,8 +1,69 @@
-body {
-  padding: 50px;
-  font: 14px "Lucida Grande", Helvetica, Arial, sans-serif;
+* {
+  font-family: "Lucida Grande", Helvetica, Arial, sans-serif;
 }
 
 a {
   color: #00B7FF;
+}
+
+footer {
+	margin-top: 50px;
+	padding: 20px;
+	text-align: center;
+}
+
+.wrapper {
+	max-width: 960px;
+	margin: 0 auto;
+}
+
+.flex {
+	display: flex;
+	flex-direction: row;	
+	justify-content: space-between;
+	align-items: center;
+}
+
+.title {
+	padding: 50px;
+	font-size: 2rem;
+	font-weight: bold;
+	text-align: center;
+}
+
+.ticker {
+	/*padding-top: 50px;*/
+}
+
+.ticker-info {
+	background-color: #d5dbe2;
+	padding: 20px;
+	border-radius: 3px;
+}
+
+.ticker-info-main {
+	width: 150px;
+	font-size: 1.2em;
+}
+
+.ticker-info-hl {
+	width: 150px; 
+	font-size: 0.8em;
+}
+
+.ticker-chart {
+	margin-top: 20px;
+	height: 620px;
+	border-radius: 3px;
+	background-color: #efefef;
+}
+
+.ticker-chart-waiter {
+	display: block;
+	position: relative;
+	width: 100%;
+	top: 50%;
+	text-align: center;
+	font-size: 1em;
+	color: #777777;
 }
