--- views/index.hbs
@@ -1,3 +1,0 @@
-<h1>{{title}}</h1>
-<p>Welcome to {{title}}</p>
-<h1>test</h1>
