--- views/layout.hbs
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>{{title}}</title>
-    <link rel='stylesheet' href='/stylesheets/style.css' />
-  </head>
-  <body>
-    <h1>Hello</h1>
-    <!-- {{body}} -->
-    <!-- {{index}} -->
-  </body>
-</html>
