+++ views/layout.pug
@@ -1,0 +1,12 @@
+doctype html
+html
+	head
+		title 
+			block title		
+		link(rel='stylesheet' href='/stylesheets/style.css')
+
+	body
+		.wrapper
+			block view
+
+			footer Made with love
